Airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas
requests
dbt-snowflake==1.9.0
dbt-core==1.9.0
apache-airflow-providers-snowflake